Fix pagination for JDG sites
All pages have a ' next ' link to view the next content, not the next page.
But contents with more than one page also got this link but it's not embedded in `single-test` class node.
So we first remove that block and we found the ' next ' content, it means it's multipage article.