fixes next_page_link detection for heise-ct
This commit fixes the next_page_link detection for heise.de/ct/
articles, where the next-link (<a> with class=next) refers to the next
article and not to the next page. Therefore i added a check, that the
link text does not contain the word "Artikel".

Without that check all consecutive articles are fetched, altough they
are new articles and not a follow-up page of the same article.

e.g.:

https://www.heise.de/ct/ausgabe/2017-6-DVB-T2-HD-startet-in-Deutschland-3636637.html