Create arxiv.org.txt (#870)
Arxiv is a digital archive used by scientists to publish pre-print scientific puiblications (typically, not yet accepted for publication in a journal or conference, ...).

No such file exists and applications dependent on fivefilters (in my case, wallabag.it) fail on arxiv. The failure is reproduced on the test page.

This is a simple version: it can probably be improved! In particular, it would be nice to remove the `blockquote` (i.e. only get the content) and check that the title is well extracted, but I was not able to.