Update mediapart.fr.txt
replace_string(<span class="screen-reader-only">): <span class="screen-reader-only" is used to make the article introduction a little bolder to distinguish it from the rest of the article, like it is in the original web render of the article

strip_id_or_class: screen-reader-only and replace_string(<p class="news__heading__top__intro): <h6 class="news__heading__top__intro" are used to remove content that is not expected to be displayed in the text render, especially the article date and the drop cap at the beginning of the article. The replace_string is a workaround to remove this drop cap because, for some reason, strip_id_or_class does not work for this specific element.