Fix extracting body on stackexchange sites (#1785)
As far as I can tell, the "post-text" classname is gone. Instead, the new
approach uses schema.org microdata instead of relying on IDs and classnames
that might change in the future.

The XPath selectors are tricky, but basically we are looking for HTML
containers that have the `itemscope` attribute and selecting child elements
based on their `itemprop` attribute.

As a bonus, this additionally extracts the question title (to avoid relying
on the <title> tag), author, and question date.